# src/build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"GenericTypeORMErrorType": {
				"description": "A generic error given by TypeORM. Often related to invalid syntax for input parameters",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "driverError: error: invalid input syntax for type uuid: \"abcd\""
				}
			},
			"requiredFieldErrorType": {
				"description": "A error that can occur when you don't give a required input parameter in the body of a request",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Must provide an expiration name when expiring a PIN"
				}
			},
			"serverErrorType": {
				"description": "Generic unknown server error",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Internal Server Error"
				}
			},
			"UserRoles": {
				"type": "string",
				"enum": [
					"Standard",
					"Admin",
					"SuperAdmin"
				]
			},
...
			"unauthorizedError": {
				"description": "Unauthorized Error",
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Unauthorized error",
					"code": 401
				}
			},
			"badRequestError": {
				"description": "Bad Request Error",
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Bad request error",
					"code": 400
				}
			},
			"forbiddenError": {
				"description": "Forbidden Error",
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Forbidden error",
					"code": 403
				}
			},
			"notFoundError": {
				"description": "Not Found Error",
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "Not found error",
					"code": 404
				}
			},
...
			"userList": {
				"description": "Response for user list",
				"properties": {
					"userId": {
						"type": "string"
					},
					"userGuid": {
						"type": "string"
					},
					"identityType": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRoles"
					},
					"organization": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"deactivationReason": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"userId",
					"userGuid",
					"identityType",
					"role",
					"organization",
					"email",
					"userName",
					"givenName",
					"lastName",
					"isActive",
					"deactivationReason",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"userId": "82dc08e5-cbca-40c2-9d35-a4d1407d5f8d",
					"userGuid": "A84D1AB221334298956C47A7B623E983",
					"identityType": "idir",
					"role": "Admin",
					"organization": "Bc Service",
					"email": "abc@gov.ca",
					"userName": "johndoe",
					"givenName": "John",
					"lastName": "Doe",
					"isActive": true,
					"deactivationReason": "To get access to site",
					"updatedAt": "2023-08-25T15:12:59.764Z"
				}
			},
			"noActiveUserFound": {
				"description": "No active user found",
				"properties": {
					"message": {
						"type": "string"
					},
					"code": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"code"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"message": "No active request found",
					"code": 204
				}
			},
			"userListQueryParam": {
				"description": "The enum is for validation to get user list query param to retrieve active or\nnon active users",
				"enum": [
					"true",
					"false"
				],
				"type": "string"
			},
			"userUpdateRequestBody": {
				"description": "Request Body for user update",
				"properties": {
					"userId": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRoles"
					},
					"organization": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"givenName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"role",
					"organization",
					"email",
					"userName",
					"givenName",
					"lastName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"userDeactivateRequestBody": {
				"description": "Request body for deactivation of one or multiple users",
				"properties": {
					"userIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"deactivationReason": {
						"type": "string"
					},
					"givenNames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"lastNames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"emails": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userIds",
					"deactivationReason",
					"givenNames",
					"lastNames",
					"emails"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"userIds": [
						"82dc08e5-cbca-40c2-9d35-a4d1407d5f8d"
					],
					"deactivationReason": "reason for deactivation",
					"givenNames": [
						"John"
					],
					"lastNames": [
						"Smith"
					],
					"emails": [
						"example@test.com"
					]
				}
			}
		},
		"securitySchemes": {
			"vhers_api_key": {
				"type": "apiKey",
				"in": "header",
				"name": "x-api-key",
				"schema": {
					"type": "string",
					"format": "uuid"
				},
				"required": true
			}
		}
	},
	"info": {
		"title": "bc-emli-pin-mgmt-be",
		"version": "1.0.0",
		"description": "BC Virtual Home Energy Rating PIN Management System Backend",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "Government of British Columbia"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/userList"
									},
									"type": "array"
								}
							}
						}
					},
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/noActiveUserFound"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/badRequestError"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/unauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/forbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/notFoundError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericTypeORMErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "active",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/userListQueryParam"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"204": {
						"description": "No content"
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/forbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/notFoundError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/requiredFieldErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/userUpdateRequestBody"
							}
						}
					}
				}
			}
		},
		"/users/deactivate": {
			"put": {
				"operationId": "DeactivateUsers",
				"responses": {
					"204": {
						"description": "No content"
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/forbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/notFoundError"
								}
							}
						}
					},
					"422": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/requiredFieldErrorType"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/serverErrorType"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/userDeactivateRequestBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}